🚚 Rename:  폴더명 오타 수정하여 반영
