+++ src/features/register/searchDepartment/queries/departmentQueries.ts
